# data-ref/src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API Baseline
  description: CRUD Operation on repositories data
  version: 0.7.0
servers:
  - url: http://data-ref.provoly.com
paths:
...
  /datasets:
    post:
      tags:
        - Dataset Controller
      summary: Save a new dataset for a specific oClass
      description: |
        A dataset is a coherent, versioned (functionally) set of data belonging to the same class.
        Dataset could be CLOSED and will therefore have several versions, OPEN or MODIFIABLE.
      operationId: com.provoly.ref.dataset.DatasetController_addDataset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetDto'
            examples:
              closedDataset:
                summary: Adding a closed dataset
                value: |
                  {
                    "id": "423b5c01-1816-41d4-b358-200000000010",
                    "name": "closed dataset",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "type": "CLOSED"
                  }
              datasetWithMetadata:
                summary: Adding a closed dataset with metadata
                value: |
                  {
                    "id": "423b5c01-1816-41d4-b358-200000000010",
                    "name": "closed dataset",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "type": "CLOSED",
                    "metadata": [
                      "value": "20",
                      "metadataDefId": "c18b19eb-0c2a-4877-a7cb-3603ba94cb1c"
                    ]
                  }
              datasetWithGroups:
                summary: Adding a closed dataset with groups
                value: |
                  {
                    "id": "423b5c01-1816-41d4-b358-200000000010",
                    "name": "closed dataset",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "type": "CLOSED",
                    "groups": [
                      "ALL",
                      "test"
                    ]
                  }
              invalidDataset:
                summary: Adding an incorrect dataset (with invalid type)
                value: |
                  {
                    "id": "423b5c01-1816-41d4-b358-200000000010",
                    "name": "closed incorrect dataset",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "type": "UNKNOWN",
                    "groups": "[group_test]"
                  }
      responses:
        "201":
          description: Created
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "409":
          description: Conflict
          content:
            application/json:
              examples:
                sameOClassName:
                  summary: Save dataset with wrong name
                  description: Occurs when you try to save a dataset that has the same name as its OClass
                  value: |
                    {
                      "application": "data-ref",
                      "code": "CONFLICT",
                      "message": "Name {{datasetName}} already exists for class {{oClassName}}"
                    }
      security:
        - oidc:
            - dataset_write
    put:
      tags:
        - Dataset Controller
      summary: Update a dataset based on its identifier
      description: You can only update the dataset name, description, metadata and groups, you cannot change its id, type or OClass.
      operationId: com.provoly.ref.dataset.DatasetController_update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetDto'
            examples:
              dataset_update:
                summary: Updating a dataset name
                value: |
                  {
                    "id": "423b5c01-1816-41d4-b358-200000000010",
                    "name": "we change dataset name",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "type": "CLOSED"
                  }
              group_update:
                summary: Update dataset groups
                description: |
                  Updating groups.
                  If you add an empty group list, the current groups in the dataset will be deleted.
                  If the group list is null, the dataset's current groups will not be modified.
                value: |
                  {
                    "id": "423b5c01-1816-41d4-b358-200000000010",
                    "name": "we change dataset name",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "type": "CLOSED",
                    "groups": [
                      "AUTHENTICATED"
                    ]
                  }
      responses:
        "200":
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupErrors'
              examples:
                deleteGroupReferencedByDashboard:
                  summary: Delete group used by a dashboard
                  description: When groups are deleted from a dataset, a list of groups is returned if they were used by a dashboard.
                  value: |
                    { "missingGroupsByEntity": 
                      {
                        "dataset_id": ["group1"],
                        "dataset_id2": ["group1", "group2"]
                      }
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
          content:
            application/json:
              examples:
                changeType:
                  summary: Update dataset with different type
                  description: Occurs when you try to update a dataset with a different type
                  value: |
                    {
                      "application": "data-ref",
                      "code": "FORBIDDEN",
                      "message": "The dataset's type is immutable"
                    }
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
    get:
      tags:
        - Dataset Controller
      summary: Get all datasets
      operationId: com.provoly.ref.dataset.DatasetController_getAll
      responses:
        "201":
          description: Retrieve all datasets
          content:
            application/json:
              examples:
                getAllDatasets:
                  summary: Get all datasets
                  value: |
                    [
                      {
                        "id": "618af062-d722-43d2-8e4d-cb65d07230ac",
                        "name": "classification des sols",
                        "oClass": "9d37e039-272b-463e-942e-6c6af0c2f440",
                        "type": "CLOSED",
                        "metadata": []
                      }
                    ]
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
        - oidc:
            - dataset_read
  /datasets/id/{id}:
    get:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_get
      summary: Retrieve a dataset based on its identifier.
      description: Also returns the metadata attached to the dataset, if any.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/datasetDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_read
            - search
    delete:
      tags:
        - Dataset Controller
      summary: Delete a dataset based on its identifier
      description: Deleting a dataset is not possible if dataset-version are associated with it
      operationId: com.provoly.ref.dataset.DatasetController_delete
      parameters:
        - name: id
          in: path
          description: dataset UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "400":
          description: Bad request
          content:
            application/json:
              examples:
                deleteDataset:
                  summary: Delete dataset with dataset-version
                  description: Occurs when you try to delete a dataset who owns dataset-version
                  value: |
                    {
                      "application": "data-ref",
                      "code": "BAD_REQUEST",
                      "message": "You're not allowed to delete dataset {{datasetId}} because it owns one or more dataset version"
                    }
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
...
  /datasets/id/{id}/dataset-versions:
    get:
      tags:
        - Dataset Controller
      summary: Retrieve all dataset versions by the identifier of its dataset
      operationId: com.provoly.ref.dataset.DatasetController_getAllById
      parameters:
        - name: id
          in: path
          description: dataset UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/listdatasetVersionDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - search
            - dataset_read
            - datasource_read
  /datasets/{id}/associations:
    get:
      tags:
        - Dataset Controller
      summary: Retrieve all usages of a dataset
      operationId: com.provoly.ref.dataset.DatasetController_getAssociationsByDatasetId
      parameters:
        - name: id
          in: path
          description: UUID of the dataset
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/AssociationsDto'
              examples:
                attributeAssociations:
                  summary: Attribute associations
                  description: The parameter usedElsewhere in the response body is a boolean used to indicate if the attribute is used in a private object.
                  value: |
                    {
                      "associations": [
                            {
                              "id": "2f021992-5fb0-42d3-9646-0affccb98101",
                              "name": "Name du widget",
                              "type": "WIDGET"
                            },
                            {
                              "id": "2f021992-5fb0-42d3-9646-0affccb98111",
                              "name": "Name of the namedquery",
                              "type": "NAMED_QUERY"
                            },
                            {
                              "id": "2f121992-5fb0-42d3-9646-0affccb98111",
                              "name": "Name of the dashboard",
                              "type": "DASHBOARD"
                            }
                      ],
                      "usedElsewhere": false
                    }
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_read
  /datasets/id/{datasetId}/metadata/id/{metadataDefId}:
    put:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_setMetadata
      summary: Add metadata to a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: metadataDefId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetadataValueWriteDto'
            examples:
              add_metadata:
                summary: Adding metadata to a dataset
                value: |
                  {
                    "value": "22"
                  }
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          $ref: '#/components/responses/wrongMetadataType'
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
    delete:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_deleteMetadata
      summary: Delete metadata of a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: metadataDefId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - dataset_write
  /datasets/class/{id}:
    get:
      tags:
        - Dataset Controller
      summary: Retrieve all datasets associated to a class
      operationId: com.provoly.ref.dataset.DatasetController_getAllForClass
      parameters:
        - name: id
          in: path
          description: Class UUID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          $ref: '#/components/responses/datasetDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          $ref: '#/components/responses/notFoundException'
      security:
        - oidc:
            - class_read
            - search
  /datasets/name/{name}:
    get:
      tags:
        - Dataset Controller
      operationId: com.provoly.ref.dataset.DatasetController_getByName
      summary: Retrieve a dataset based on its name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: '#/components/responses/datasetDetailsDto'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "404":
          description: Not Found
          content:
            application/json:
              examples:
                notFound:
                  description: Occurs when you try to find a dataset with the wrong name
                  value: |
                    {
                      "application": "data-ref",
                      "code": "NOT_FOUND",
                      "message": "Dataset {{datasetName}} doesn't exists"
                    }
      security:
        - oidc:
            - dataset_read
...
components:
  responses:
    notFoundException:
      description: Entity not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    wrongMetadataType:
      description: Not Allowed
      content:
        application/json:
          examples:
            wrongMetadataType:
              summary: Add metadata with unknown type
              description: Occurs when you try to add metadata with a type not supported
              value: |
                {
                  "application": "data-ref",
                  "code": "FORBIDDEN",
                  "message": "Metadata value should be of type : {{type}}"
                }
    datasetDetailsDto:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DatasetDetailsDto'
          examples:
            dataset:
              summary: Get a dataset with metadata
              value: |
                {
                  "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "oClass": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "type": "OPEN",
                  "name": "open dataset of my class",
                  "metadata": [
                    {
                      "value": 2,
                      "metadataDef": {
                        "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "name": "metadata_name",
                        "type": "INTEGER",
                        "description": "description of metadata"
                      }
                    }
                  ]
                }
    listdatasetVersionDetailsDto:
      description: OK
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/DatasetVersionDetailsDto'
          examples:
            getAllDatasetVersion:
              summary: Get list of dataset version disabled with warnings
              description: Example of a disabled dataset version that has warnings.
              value: |
                [
                  {
                    "id": "b1a08d9d-4c38-4d43-90f3-25015537ebfd" ,
                    "dataset": "423b5c01-1816-41d4-b358-200000000010",
                    "oClass": "991eb4b2-253c-4875-b708-8209fa5bfd51",
                    "lastModified": "2023-04-27T13:56:13.861018Z",
                    "version": 1,
                    "state": "INACTIVE",
                    "withFile": false,
                    "hasWarnings": true
                  }
                ]
            datasetVersionWithMetadata:
              summary: Get list of dataset version with metadata
              description: Example of a dataset version with metadata
              value: |
                [
                  {
                    "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "dataset": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "oClass": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "lastModified": "2022-03-10T16:15:50Z",
                    "version": 0,
                    "state": "LOADING",
                    "withFile": true,
                    "metadata": [
                      {
                        "value": 2,
                        "metadataDef": {
                          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                          "name": "metadata_name",
                          "type": "INTEGER",
                          "description": "description of metadata",
                          "slug": "string"
                        }
                      }
                    ]
                  }
                ]
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: 404
        message:
          type: string
          example: "Entity : ac2c1a7f-0694-4f03-bb00-d349adeecd3d inexistant."
      required:
        - code
        - message
...
    AssociationsDto:
      type: object
      properties:
        associations:
          $ref: '#/components/schemas/AssociationDto'
        usedElsewhere:
          type: boolean
          default: false
    AssociationDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "oclass name"
          description: Name of the associated entity
        type:
          enum:
            - ABAC
            - LINK
            - NAMED_QUERY
            - DATASET
            - OCLASS
            - DASHBOARD
            - WIDGET
          example: OCLASS
...
    DatasetDetailsDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my dataset"
        oClass:
          $ref: '#/components/schemas/UUID'
        type:
          $ref: '#/components/schemas/DatasetType'
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/MetadataValueReadDto'
        groups:
          type: array
          items:
            type: string
          example: [ "first_group", "second_group" ]
        description:
          type: string
          example: "my description"
        owner:
          type: boolean
          example: false
        slug:
          type: string
    DatasetDto:
      type: object
      required:
        - id
        - name
        - oClass
        - type
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my dataset"
        oClass:
          $ref: '#/components/schemas/UUID'
        type:
          $ref: '#/components/schemas/DatasetType'
        description:
          type: string
          example: "my description"
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/MetadataValueWriteDto'
        groups:
          type: array
          items:
            type: string
          example: ["first_group", "second_group"]
    DatasetState:
      enum:
        - LOADING
        - INDEXING
        - ACTIVE
        - ERROR
        - INACTIVE
      type: string
      example: ACTIVE
    DatasetType:
      enum:
        - OPEN
        - CLOSED
        - MODIFIABLE
      type: string
      example: CLOSED
    DatasetVersionDetailsDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        dataset:
          $ref: '#/components/schemas/UUID'
        oClass:
          $ref: '#/components/schemas/UUID'
        lastModified:
          $ref: '#/components/schemas/Instant'
        version:
          format: int32
          type: integer
          example: 2
        state:
          $ref: '#/components/schemas/DatasetState'
        withFile:
          type: boolean
          example: false
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/MetadataValueReadDto'
        hasWarnings:
          type: boolean
          example: false
...
    GroupErrors:
      type : object
      additionalProperties:
        description: missing group list
        example: ["admin"]
        type: array
        items:
          type: string
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...
    MetadataDefDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "my metadata definition"
        type:
          $ref: '#/components/schemas/VariableTypeDto'
        description:
          type: string
          example: "metadata definition description"
        allowedValues:
          type: array
          items:
            type: string
        slug:
          type: string
          example: "87280_my_metadata_definition"
        readOnly:
          type: boolean
        system:
          type: boolean
    MetadataValueReadDto:
      type: object
      properties:
        value:
          type: string
          example: "my_value"
        metadataDef:
          $ref: '#/components/schemas/MetadataDefDto'
    MetadataValueWriteDto:
      type: object
      properties:
        value:
          type: string
        metadataDefId:
          $ref: '#/components/schemas/UUID'
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    VariableTypeDto:
      enum:
        - INTEGER
        - STRING
        - DOUBLE
        - DATE
        - UUID
        - LIST
      type: string
...
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://auth/realms/provoly/.well-known/openid-configuration